rework tag components into TagForm, make admin presentable (TAGFORM STILL NOT FUNCTIONAL)
