+++ src/redux/sagas/admin.saga.js
@@ -8,8 +8,6 @@ function* postResource(action) {
     try {
         //post request with new resource
         yield call(axios.post, '/api/admin', action.payload);
-        //dispatch to fetch updated list
-        yield put({ type: 'FETCH_SEARCH' });
     } catch (error) {
         console.log('Error adding resource', error);
     }
@@ -21,8 +19,6 @@ function* putResource(action) {
     try {
         //sends a PUT request to the server to update a resource
         yield call(axios.put, `/api/admin/${action.payload.id}`, action.payload);
-        //dispatches a fetch to get the updated list
-        yield put({ type: 'FETCH_SEARCH' });
     } catch (error) {
         console.log('Error updating resource', error);
     }
@@ -34,8 +30,6 @@ function* deleteResource(action) {
     try {
         //sends a DELETE request to the server to delete a resource
         yield call(axios.delete, `/api/admin/${action.payload}`);
-        //dispatches a fetch to get list without deleted resource
-        yield put({ type: 'FETCH_SEARCH' });
     } catch (error) {
         console.log('Error deleting resource', error);
     }
